+++ packages/namespaces-react/src/useNamespace.js
@@ -7,9 +7,9 @@ import { NamespaceContext } from './contexts';
 
 const useNamespace = feature => {
 	const { namespaces = {}, useNamespace = alwaysNull } = useContext(NamespaceContext);
-	const namespace = useNamespace(feature || DEFAULT_FEATURE, namespaces);
+	const namespace = useNamespace(feature ?? DEFAULT_FEATURE, namespaces);
 
-	return namespaces[feature] || namespace || namespaces[DEFAULT_FEATURE];
+	return namespaces[feature] ?? namespace ?? null;
 };
 
 export default useNamespace;